requirements.txt
black
celery
dj-database-url
Django
django-admin-rangefilter
django-celery-beat
django-dbbackup
django-debug-toolbar
django-extensions
django-filter
django-localflavor
django-logentry-admin
gunicorn
jsondate
psycopg2-binary
python-decouple
pytz
raven
redis
